+++ src/test_all.zig
@@ -14,6 +14,7 @@ test {
     _ = @import("quic/crypto_stream.zig");
     _ = @import("quic/packet_packer.zig");
     _ = @import("quic/tls13.zig");
+    _ = @import("quic/rsa.zig");
     _ = @import("quic/mtu.zig");
     _ = @import("quic/stateless_reset.zig");
     _ = @import("quic/connection_manager.zig");